+++ packages/core/package.json
@@ -1,6 +1,5 @@
 {
-  "name": "@infinity-keys/utils",
-  "private": true,
+  "name": "@infinity-keys/core",
   "packageManager": "yarn@3.2.3",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
@@ -15,6 +14,7 @@
     "cloudinary-build-url": "^0.2.4",
     "jose": "^4.11.0",
     "nanoid": "^4.0.0",
-    "uuid": "^9.0.0"
+    "uuid": "^9.0.0",
+    "zod": "^3.19.1"
   }
 }
