+++ packages/react-lens-share-button/package.json
@@ -1,6 +1,5 @@
 {
   "name": "@infinity-keys/react-lens-share-button",
-  "private": false,
   "version": "0.0.4",
   "type": "module",
   "scripts": {
