--- packages/types/README.md
@@ -1,1 +1,0 @@
-# types
