--- packages/types/package.json
@@ -1,18 +1,0 @@
-{
-  "name": "@infinity-keys/types",
-  "private": true,
-  "version": "0.0.0",
-  "packageManager": "yarn@3.2.3",
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "scripts": {
-    "build": "tsc"
-  },
-  "devDependencies": {
-    "typescript": "^4.8.4"
-  },
-  "dependencies": {
-    "jose": "^4.11.0",
-    "zod": "^3.19.1"
-  }
-}
