+++ web/package.json
@@ -18,9 +18,8 @@
     "@heroicons/react": "^2.0.11",
     "@infinity-keys/constants": "workspace:^",
     "@infinity-keys/contracts": "workspace:^",
+    "@infinity-keys/core": "workspace:^",
     "@infinity-keys/react-lens-share-button": "workspace:^",
-    "@infinity-keys/types": "workspace:^",
-    "@infinity-keys/utils": "workspace:^",
     "@rainbow-me/rainbowkit": "^0.7.0",
     "@redwoodjs/forms": "^3.2.0",
     "@redwoodjs/router": "^3.2.0",
